Backend/requirement.txt
SQLAlchemy[asyncio]
psycopg2-binary
python-dotenv
mcp
trio
fastapi
uvicorn
typer
rich
requests
icalendar